AssetsUSA/requirements.txt
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
Django==5.1.7
django-storages==1.14.5
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0